Reset read receipt on sending sticker
